Add provider trait for CanSendJsonRpcRequest
