+++ relayer/crates/starknet-chain-components/src/traits/mod.rs
@@ -1,14 +1,10 @@
 pub mod account;
 pub mod client;
 pub mod contract;
+pub mod json_rpc;
 pub mod messages;
 pub mod proof_signer;
 pub mod queries;
+pub mod rpc_client;
 pub mod transfer;
 pub mod types;
-
-mod json_rpc;
-mod rpc_client;
-
-pub use json_rpc::*;
-pub use rpc_client::*;
